+++ src/translators/index.ts
@@ -4,7 +4,6 @@ export * from './FakeTranslator';
 
 // Stable translators
 export * from './MicrosoftTranslator';
-export * from './YandexTranslator';
 export * from './GoogleTranslator';
 export * from './DeepLTranslator';
 export * from './TartuNLPTranslator';
